requirements.txt

streamlit==1.28.1
groq==0.4.1
crewai==0.28.8
langchain==0.1.5
langchain-community==0.0.13
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0